move UaDateTime into own cpp file
